requirements.txt
# --- Web API / Server ---
fastapi==0.115.2
uvicorn[standard]==0.30.6
jinja2==3.1.4

# --- IA / NLP / Vision ---
transformers==4.43.4
diffusers==0.29.2
accelerate==0.34.2
safetensors==0.4.4
sentencepiece==0.2.0
protobuf==5.27.3
Pillow==10.4.0

# --- RAG (TF-IDF) ---
scikit-learn==1.5.2
numpy==1.26.4

# --- Utilidades lingüísticas (opcional) ---
nltk==3.9.1

# --- Gráficos educativos (Manim) ---
manim==0.18.1

# --- Compatibilidad tipado ---
typing-extensions==4.12.2
